requirements.txt
pandas>=1.5.0
numpy>=1.21.0
kafka-python>=2.0.2
matplotlib>=3.5.0
seaborn>=0.11.0 